Working :D

+++ src/index.ts
@@ -4,16 +4,21 @@ import {captureRegionsReturnMat} from "./RegionCapture";
 import {parseWords} from "./parseWords";
 import {parseBoard} from "./parseBoard";
 import {searchWords} from "./searchWords";
+import {TestCapture} from "./TestCapture";
+import {CompleteGame} from "./Interactions";
 
+let ENV = "prod"
 
 const Main = async (width: number, height: number) => {
-    const {boardMat, wordsMat} = captureRegionsReturnMat(width, height);
+    const {boardMat, wordsMat} = ENV === "prod" ? captureRegionsReturnMat(width, height) : TestCapture();
     const words = await parseWords(wordsMat);
     const board = await parseBoard(boardMat);
     if (!board || !words)
         return printAndExit('Failed to parse board or words', 3);
-    console.log(words)
-    searchWords(words, board)
+    const Positions = searchWords(words, board)
+    if (!Positions)
+        return printAndExit('Failed to search words', 4);
+    await CompleteGame(Positions)
     console.log("yey")
     return;
 };
@@ -25,8 +30,12 @@ const Main = async (width: number, height: number) => {
     if (!config.width || !config.height)
         printAndExit('Config file must contain width and height', 2);
 
-    fs.readdirSync('./img').forEach((file) => {
-        fs.unlinkSync(`./img/${file}`);
-    });
+    ENV = process.argv[2] === 'dev' ? 'dev' : 'prod';
+
+    if (ENV === 'prod') {
+        fs.readdirSync('./img').forEach((file) => {
+            fs.unlinkSync(`./img/${file}`);
+        });
+    }
     await Main(config.width, config.height);
 })();
